fw_config: Convert fw_config to a 64-bit field
We all knew this was coming, 32 bits is never enough. Doing this early
so that it doesn't affect too much code yet. Take care of every usage of
fw_config throughout the codebase so the conversion is all done at once.

BUG=b:169668368
TEST=Hacked up this code to OR 0x1_000_0000 with CBI-sourced FW_CONFIG
and verify the console print contained that bit.

Reviewed-on: https://review.coreboot.org/c/coreboot/+/45939
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>

+++ src/ec/google/chromeec/ec.h
@@ -83,7 +83,7 @@ int google_chromeec_reboot(int dev_idx, enum ec_reboot_cmd type, uint8_t flags);
  */
 int google_chromeec_cbi_get_oem_id(uint32_t *id);
 int google_chromeec_cbi_get_sku_id(uint32_t *id);
-int google_chromeec_cbi_get_fw_config(uint32_t *fw_config);
+int google_chromeec_cbi_get_fw_config(uint64_t *fw_config);
 int google_chromeec_cbi_get_dram_part_num(char *buf, size_t bufsize);
 int google_chromeec_cbi_get_oem_name(char *buf, size_t bufsize);
 /* version may be stored in CBI as a smaller integer width, but the EC code
